Do some package reorganization.
